backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
opencv-python==4.8.1.78
mediapipe==0.10.8
tensorflow==2.15.0
numpy==1.24.3
Pillow==10.1.0
pandas==2.1.3
tqdm==4.66.1
scikit-learn==1.3.2
matplotlib==3.8.2